[EBPF] Use uint64 instead of uintptr in CGo types (#29293)

+++ pkg/network/ebpf/kprobe_types_linux.go
@@ -77,12 +77,12 @@ type PIDFD struct {
 	Fd  uint32
 }
 type UDPRecvSock struct {
-	Sk  uintptr
-	Msg uintptr
+	Sk  uint64
+	Msg uint64
 }
 type BindSyscallArgs struct {
-	Addr uintptr
-	Sk   uintptr
+	Addr uint64
+	Sk   uint64
 }
 type ProtocolStack struct {
 	Api         uint8
